Refactor : 지도 마커 생성과 지도 그리는 부분 로직 분리 - 로직을 분리해서 마커 생성해주는 컴포넌트로 props map를 넘기도록 함..( 이게 아닌 것 같음.. )
